+++ UC_Errors.py
@@ -1,0 +1,2 @@
+class UnitError(Exception): pass
+class FileFormatError(Exception): pass
