+++ test.py
@@ -1,4 +1,4 @@
-from Unit import *
+from UC_Unit import *
 
 def test_fail(msg, verbose):
 	if verbose: print(f"Test failed: {msg}")
